Clean up BootstrapperTest
